add -Wconversion option

+++ src/cr_signal.cc
@@ -19,7 +19,7 @@ extern "C" {
 namespace cris::core {
 
 static void WriteToGlog(const char* data, int size) {
-    std::string msg(data, size);
+    std::string msg(data, static_cast<std::size_t>(size));
     LOG(ERROR) << msg;
 }
